+++ R/corPlot.R
@@ -123,7 +123,7 @@ corPlot <- function(
   r.thresh = 0.8,
   text.col = c("black", "black"),
   key = FALSE,
-  key.header = NULL,
+  key.title = NULL,
   key.position = "right",
   strip.position = "top",
   auto.text = TRUE,
@@ -142,6 +142,9 @@ corPlot <- function(
 
   # extra.args setup
   extra.args <- list(...)
+
+  # check if key.header / key.footer are being used
+  key.title <- check_key_header(key.title, extra.args)
 
   # deprecated lower arg
   if ("lower" %in% names(extra.args)) {
@@ -476,7 +479,7 @@ corPlot <- function(
       x = extra.args$xlab,
       y = extra.args$ylab,
       title = extra.args$main,
-      fill = key.header
+      fill = key.title
     )
 
   # if dendrogram, need to use legendry to switch dendro to the opposite side of